+++ provider/internal/host.go
@@ -96,10 +96,16 @@ func (h *host) builderFor(ctx context.Context, build Build) (*cachedBuilder, err
 		builder.WithStore(txn),
 	)
 	if err != nil && build.ShouldExec() && strings.HasPrefix(opts.Builder, "cloud-") {
-		err = errors.Join(err, errors.New("Make sure you're logged in to Docker (`docker login`) if you're trying to use a cloud builder."))
+		//nolint:revive // Human-readable.
+		err = errors.Join(err, errors.New(
+			"Make sure you're logged in to Docker (`docker login`) if you're trying to use a cloud builder."),
+		)
 	}
 	if err != nil && build.ShouldExec() {
-		err = errors.Join(err, errors.New("Make sure your buildx plugin is executable (`docker buildx version`)"))
+		//nolint:revive // Human-readable.
+		err = errors.Join(err, errors.New(
+			"Make sure your buildx plugin is executable (`docker buildx version`)"),
+		)
 	}
 	if err != nil {
 		return nil, fmt.Errorf("new builder: %w", err)
@@ -168,7 +174,10 @@ func (h *host) builderFor(ctx context.Context, build Build) (*cachedBuilder, err
 	nodes, err := b.LoadNodes(ctx, builder.WithData())
 	if err != nil && !build.ShouldExec() {
 		if strings.Contains(err.Error(), "failed to find driver") {
-			err = errors.Join(err, errors.New("Use `exec: true` if you're trying to use Docker Build Cloud or other custom drivers."))
+			//nolint:revive // Human-readable.
+			err = errors.Join(err, errors.New(
+				"Use `exec: true` if you're trying to use Docker Build Cloud or other custom drivers.",
+			))
 		}
 		return nil, fmt.Errorf("loading nodes: %w", err)
 	}